TST/REF: misplaced reindex/reset_index tests
